Refactor Access Modes management
Access modes are now part of the shared code rather than the gConsent one. They also support the short version of modes the JSON-LD document may contain
