+++ src/index.ts
@@ -32,7 +32,7 @@ export type {
 
 export type { FetchOptions } from "./type/FetchOptions";
 export type { RedirectOptions } from "./type/RedirectOptions";
-export type { AccessModes, SerializedAccessModes } from "./type/AccessModes";
+export type { AccessModes } from "./type/AccessModes";
 export type { AccessGrantAny } from "./type/AccessGrant";
 
 export {
