+++ src/type/ResourceAccessMode.ts
@@ -19,7 +19,22 @@
 // SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-import type { RESOURCE_ACCESS_MODE } from "../constants";
+export const ACL_RESOURCE_ACCESS_MODE_APPEND =
+  "http://www.w3.org/ns/auth/acl#Append";
+export const ACL_RESOURCE_ACCESS_MODE_READ =
+  "http://www.w3.org/ns/auth/acl#Read";
+export const ACL_RESOURCE_ACCESS_MODE_WRITE =
+  "http://www.w3.org/ns/auth/acl#Write";
+
+export const RESOURCE_ACCESS_MODE = new Set([
+  ACL_RESOURCE_ACCESS_MODE_APPEND,
+  ACL_RESOURCE_ACCESS_MODE_READ,
+  ACL_RESOURCE_ACCESS_MODE_WRITE,
+  // The following are linked to the previous through our JSON-LD context.
+  "Read",
+  "Write",
+  "Append",
+] as const);
 
 export type ResourceAccessMode = typeof RESOURCE_ACCESS_MODE extends Set<
   infer T
